BlockHandler.update() returns a boolean
Indicating that the block should or should not continue to the next block during the frame. Otherwise the flow will be done updating for that frame. This is to allow event-based blocks like keyrelease to only be called once per input frame.

Also, added a cap to how fast objects can travel. Experimental.